+++ phantom-dsl/src/main/scala/com/websudos/phantom/builder/syntax/CQLSyntax.scala
@@ -151,6 +151,7 @@ object CQLSyntax {
   object Selection {
     val BlobAsText = "blobAsText"
     val DateOf = "dateOf"
+    val UnixTimestampOf = "unixTimestampOf"
     val OrderBy = "ORDER BY"
     val MaxTimeUUID = "maxTimeuuid"
     val MinTimeUUID = "minTimeuuid"